refactor: rename ySweetSync → ySweetPersistSync, require persistence
Clean break rename (no backward-compat shims):
- ySweetSync → ySweetPersistSync, YSweetSyncConfig → YSweetPersistSyncConfig
- persistence config field is now required (was optional)
- Move persistence/web.ts and persistence/desktop.ts into y-sweet-persist-sync/
- Delete extensions/persistence/ directory and conditional exports
- Update all call sites in tab-manager
- Update package.json exports
- Rename and update test file (3/3 pass)

Spec: specs/20260213T104500-y-sweet-persist-sync-rename.md
