+++ packages/epicenter/package.json
@@ -12,13 +12,10 @@
 			"node": "./src/extensions/index.ts",
 			"default": "./src/extensions/index.ts"
 		},
-		"./extensions/persistence": {
-			"browser": "./src/extensions/persistence/index.browser.ts",
-			"node": "./src/extensions/persistence/index.node.ts",
-			"default": "./src/extensions/persistence/index.node.ts"
-		},
+		"./extensions/y-sweet-persist-sync": "./src/extensions/y-sweet-persist-sync.ts",
+		"./extensions/y-sweet-persist-sync/web": "./src/extensions/y-sweet-persist-sync/web.ts",
+		"./extensions/y-sweet-persist-sync/desktop": "./src/extensions/y-sweet-persist-sync/desktop.ts",
 		"./extensions/websocket-sync": "./src/extensions/websocket-sync.ts",
-		"./extensions/y-sweet-sync": "./src/extensions/y-sweet-sync.ts",
 		"./cli": "./src/cli/index.ts",
 		"./server": "./src/server/index.ts",
 		"./ingest": "./src/ingest/index.ts",
