--- packages/epicenter/src/extensions/persistence/index.browser.ts
@@ -1,8 +1,0 @@
-/**
- * Browser persistence provider entry point.
- * Uses IndexedDB via y-indexeddb for YJS document persistence.
- *
- * This file is selected by bundlers when the "browser" condition is matched
- * in package.json exports.
- */
-export { indexeddbPersistence } from './web.js';
